+++ packages/compile/index.js
@@ -2,16 +2,28 @@ module.exports = (Goat) => new Goat({
   name: 'Compile',
   command: 'compile',
   description: 'Webpack based compiler for Javascript and scss',
-  schema: require('./scripts/schema'),
-  options: {
-    '-p, --production': 'Production',
-  },
+  schema: require('./schema/schema.js'),
+  options: [
+    {
+      label: 'Compile for production',
+      flags: '-p, --production',
+      allowOnOnce: true,
+      allowOnWatch: true,
+    },
+    {
+      label: 'Analyse javascript dependency sizes',
+      flags: '-a, --analyse',
+      allowOnOnce: true,
+      allowOnWatch: false,
+    },
+  ],
   async method(config) {
     const { runAll } = require('./scripts/webpack/runner');
     runAll({
       ...config,
       settings: {
         production: config.options.production,
+        analyse: config.options.analyse,
       },
     });
   },
@@ -23,5 +35,8 @@ module.exports = (Goat) => new Goat({
         production: config.options.production,
       },
     });
+  },
+  init: {
+    files: () => require('./init/files.js'),
   },
 });