+++ packages/compile/init/files/webpack.js
@@ -1,0 +1,10 @@
+module.exports.webpack = (config) => {
+  return config;
+};
+
+module.exports.transform = {
+  lodash: {
+    transform: 'lodash/${member}',
+    preventFullImport: true,
+  },
+};