+++ packages/compile/scripts/webpack/config/plugins.js
@@ -5,7 +5,6 @@
  */
 function getPlugins(config) {
   const { ProgressPlugin } = require('webpack');
-  const { CleanWebpackPlugin } = require('clean-webpack-plugin');
   const MiniCssExtractPlugin = require('mini-css-extract-plugin');
   const FixStyleOnlyEntriesPlugin = require('webpack-fix-style-only-entries');
   const FriendlyErrorsWebpackPlugin = require('friendly-errors-webpack-plugin');
@@ -22,10 +21,26 @@ function getPlugins(config) {
     }),
   ];
 
+  if (config.configuration.bundler.js.esm) {
+    const BabelEsmPlugin = require('babel-esm-plugin');
+    plugins.push(new BabelEsmPlugin({
+      filename: '[name].esm.js',
+      chunkFilename: '[id].esm.js',
+    }));
+  }
+
   if (config.configuration.bundler.clean) {
+    const { CleanWebpackPlugin } = require('clean-webpack-plugin');
     const { dist } = config.configuration.locations;
     plugins.push(new CleanWebpackPlugin({
       cleanOnceBeforeBuildPatterns: [`${dist}/**/*`],
+    }));
+  }
+
+  if (config.settings.analyse) {
+    const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
+    plugins.push(new BundleAnalyzerPlugin({
+      excludeAssets: (assetName) => assetName.includes('.esm.js'),
     }));
   }