+++ packages/compile/scripts/webpack/webpack.js
@@ -19,6 +19,19 @@ function getWebpackDevelopmentSetup() {
   };
 }
 
+function modifyWebpackConfig(config, instance) {
+  try {
+    // eslint-disable-next-line import/no-dynamic-require
+    const modifyWebpack = require(`${config.path}/.goat/webpack`).webpack(instance);
+    if (!modifyWebpack || (Object.keys(modifyWebpack).length === 0 && modifyWebpack.constructor === Object)) {
+      return instance;
+    }
+    return modifyWebpack;
+  } catch (e) {
+    return instance;
+  }
+}
+
 /**
  * Get the Webpack setup
  * @param {Object} config
@@ -29,7 +42,8 @@ function getWebpackSetup(config) {
   const common = require('./config/webpack.common')(config);
   const webpack = require('webpack');
   const setup = config.settings.production ? getWebpackProductionSetup() : getWebpackDevelopmentSetup();
-  return webpack(merge(common, setup));
+  const instance = modifyWebpackConfig(config, merge(common, setup));
+  return webpack(instance);
 }
 
 module.exports = getWebpackSetup;