+++ packages/goat/src/methods/modules/loadModule.js
@@ -8,13 +8,18 @@ const Notifier = require('../notifications/notifyHandler');
  */
 function loadModule(module) {
   if (!module.global) {
-    return require(module.package)
+    try {
+      return require(module.package);
+    } catch (error) {
+      Notifier.error(`The module ${module.package} could not be loaded, please try updating goat or remove the module from your project config.`);
+      process.exit();
+    }
   }
   try {
     return importGlobal(module.package);
-  } catch(error) {
-    Notifier.error(`The module ${module.package} doesn't seem to be installed on your system`)
-    Notifier.log(`Please install by running: ${Notifier.script(`goat module add ${module.package}`)}`)
+  } catch (error) {
+    Notifier.error(`The module ${module.package} doesn't seem to be installed on your system`);
+    Notifier.log(`Please install by running: ${Notifier.script(`goat module add ${module.package}`)}`);
     process.exit();
   }
 }