+++ packages/storybook-helpers/scripts/createAttributes.js
@@ -1,0 +1,5 @@
+function createAttributes(attributes) {
+  return Object.entries(attributes).reduce((result, { 0: attribute, 1: value }) => (`${result} ${attribute}="${value}";`), '');
+}
+
+module.exports = createAttributes;