+++ packages/storybook-helpers/scripts/createImage.js
@@ -1,0 +1,10 @@
+const { createAttributes } = require('../index');
+const image = require('./components/image.twig');
+
+function createImage(attributes) {
+  return image({
+    attributes: createAttributes(attributes),
+  });
+}
+
+module.exports = createImage;