+++ packages/storybook-helpers/scripts/drupalAttributes.js
@@ -1,0 +1,3 @@
+const DrupalAttributes = require('drupal-attribute');
+
+module.exports = DrupalAttributes;