+++ packages/storybook-helpers/scripts/faker.js
@@ -1,0 +1,3 @@
+const faker = require('faker');
+
+module.exports = faker;