+++ packages/storybook/scripts/storybook/main.js
@@ -7,7 +7,6 @@ module.exports = {
   addons: [
     '@storybook/addon-a11y',
     '@storybook/addon-links',
-    '@storybook/addon-design-assets',
     '@storybook/addon-essentials',
     '@storybook/addon-knobs',
   ],