+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 1.8.5
+### Compile
+- Feat: Added BabelEsmPlugin to generate modern javascript files - this feature needs to be enabled manually due to possible bugs regarding css, set bundler.js.esm to true in goat.config.js
+- Feat: Added babel-plugin-transform-imports to improve treeshaking. You can configure this per project via webpack.js inside the .goat folder. The lodash definition is pre loaded.
+- Feat: Allow modifing the webpack setup per project from .goat/webpack.js
+- Fix: fixes filenames of compile module, remove .es6 from compiled file
+
+## 1.8.4
+- Feat: update project config when keys are missing
+- Feat: Build and update project config using json schemas instead of configuration sample files
+
+## 1.8.3
+- Fix: various bugfixes for Compile
+- Fix: various bugfixes for Storybook
+- Fix: allow flags on watch tasks
+
+## 1.8.2
+- Feat: Compile task added based on webpack
+- Feat: Storybook task added
+- Feat: Allow for usage of flags in modules
+
 ## 1.7.2
 - Fix: Polyfill JS classes using js bundler
 
@@ -25,7 +46,6 @@
 
 ## 1.6.0
 - Added new module: JS-bundler
-
 
 ## 1.5.1
 - Feature: added `project ls` command